requirements.txt
grpcio
grpcio-tools
pandas
scikit-learn==1.5.2
joblib
uvloop
grpcio-reflection